@@ -4,7 +4,7 @@ description: A sample command-line application.
 # homepage: https://www.example.com
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   dart_numerics: ^0.0.5
@@ -14,3 +14,74 @@ dependencies:
 dev_dependencies:
   pedantic: ^1.9.0
   test: ^1.14.4
+
+
+#
+#name: genetic_algorithms
+#description: A new Flutter project.
+#
+#publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+#version: 1.0.0+1
+#
+#environment:
+#  sdk: ">=2.12.0 <3.0.0"
+#
+#dependencies:
+#  flutter:
+#    sdk: flutter
+#
+#  # logger
+#  logger: ^1.0.0
+#
+#  # bloc
+#  bloc: ^7.0.0
+#  flutter_bloc: ^7.0.0
+#
+#  # sqflite
+#  sqlite3: ^1.0.0
+#  sqlite3_flutter_libs: ^0.4.1
+#
+#  sqflite_common_ffi: ^2.0.0
+#  sqflite: ^2.0.0+3
+#  path: ^1.8.0
+#  path_provider: ^2.0.1
+#
+#  cupertino_icons: ^1.0.0
+#
+#dev_dependencies:
+#  flutter_test:
+#    sdk: flutter
+#
+#flutter:
+#  uses-material-design: true
+
+  # To add assets to your application, add an assets section, like this:
+  # assets:
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
+
+  # An image asset can refer to one or more resolution-specific "variants", see
+  # https://flutter.dev/assets-and-images/#resolution-aware.
+
+  # For details regarding adding assets from package dependencies, see
+  # https://flutter.dev/assets-and-images/#from-packages
+
+  # To add custom fonts to your application, add a fonts section here,
+  # in this "flutter" section. Each entry in this list should have a
+  # "family" key with the font family name, and a "fonts" key with a
+  # list giving the asset and other descriptors for the font. For
+  # example:
+  # fonts:
+  #   - family: Schyler
+  #     fonts:
+  #       - asset: fonts/Schyler-Regular.ttf
+  #       - asset: fonts/Schyler-Italic.ttf
+  #         style: italic
+  #   - family: Trajan Pro
+  #     fonts:
+  #       - asset: fonts/TrajanPro.ttf
+  #       - asset: fonts/TrajanPro_Bold.ttf
+  #         weight: 700
+  #
+  # For details regarding fonts from package dependencies,
+  # see https://flutter.dev/custom-fonts/#from-packages
